Handle invalid row versions
There is a weird case in READ UNCOMMITTED and READ COMMITTED where a
transaction can UPDATE/DELETE a row created by a transaction that
started later.  In effect, system-versioned tables must be run at the
SERIALIZABLE level and so if we come across such an anomaly, we give an
invalid row version error, per spec.