Require tables with FOR PORTION OF views to have a primary key.
Before this, we would find the row in the underlying table by matching
all of the columns, but this doesn't work if one of the columns doesn't
have an equality operator.

Case in point (no pun intended): the point type has no operator "=".  It
has an equality operator called "~=" but there is no way to figure that
out because point has no btree/hash opclass.

We could simply limit our tests to columns with a btree/hash opclass,
but that could produce unpredictable results on a table with no primary
key, so let's just require one.  It's good practice anyway.

Note: We don't need the null tests anymore now that we only search on
primary key.