Correctly insert into history table.
The table with system versioning might not have the same physical
structure as the history table.  For example a column may have been
deleted in the main table before the history table is created.  So, just
as with partitioning and some other cases, we have to convert the tuple
from one format to the other while copying.

Fixes #5 https://github.com/xocolatl/periods/issues/5