Show how to ALTER a table with SYSTEM VERSIONING
It is a little confusing how to make alterations to a table with system
versioning, and the fact that we don't prevent much DDL (just drops,
really) makes it even more confusing.

Add a section to the README showing how it is supposed to work.

See #6. https://github.com/xocolatl/periods/issues/6