Use main table's tupledesc if there is no map.
This corrects for the fact that PostgreSQL does not detect the need for
a map when atthasmissing is different.  The consequence of this is that
nulls are inserted as the start time in the history table for any rows
that existed in the main table before adding system_time columns.

Only versions 11+ are affected because that is when missing attributes
were first introduced.